+++ kitsune/src/service/oauth2/issuer.rs
@@ -1,0 +1,202 @@
+use super::TOKEN_VALID_DURATION;
+use crate::{error::Error, util::generate_secret};
+use async_trait::async_trait;
+use diesel::{ExpressionMethods, OptionalExtension, QueryDsl, SelectableHelper};
+use diesel_async::{scoped_futures::ScopedFutureExt, AsyncConnection, RunQueryDsl};
+use kitsune_db::{
+    model::oauth2,
+    schema::{oauth2_access_tokens, oauth2_applications, oauth2_refresh_tokens},
+    PgPool,
+};
+use oxide_auth::primitives::{
+    grant::{Extensions, Grant},
+    issuer::{RefreshedToken, TokenType},
+    prelude::IssuedToken,
+};
+use oxide_auth_async::primitives::Issuer;
+use time::OffsetDateTime;
+
+#[derive(Clone)]
+pub struct OAuthIssuer {
+    pub db_pool: PgPool,
+}
+
+#[async_trait]
+impl Issuer for OAuthIssuer {
+    async fn issue(&mut self, grant: Grant) -> Result<IssuedToken, ()> {
+        let application_id = grant.client_id.parse().map_err(|_| ())?;
+        let user_id = grant.owner_id.parse().map_err(|_| ())?;
+        let scopes = grant.scope.to_string();
+        let expired_at = OffsetDateTime::from_unix_timestamp(grant.until.timestamp())
+            .unwrap()
+            .replace_nanosecond(grant.until.timestamp_subsec_nanos())
+            .unwrap();
+
+        let mut db_conn = self.db_pool.get().await.map_err(|_| ())?;
+        let (access_token, refresh_token) = db_conn
+            .transaction(|tx| {
+                async move {
+                    let access_token = diesel::insert_into(oauth2_access_tokens::table)
+                        .values(oauth2::NewAccessToken {
+                            token: generate_secret().as_str(),
+                            user_id: Some(user_id),
+                            application_id: Some(application_id),
+                            scopes: scopes.as_str(),
+                            expired_at,
+                        })
+                        .returning(oauth2::AccessToken::as_returning())
+                        .get_result::<oauth2::AccessToken>(tx)
+                        .await?;
+
+                    let refresh_token = diesel::insert_into(oauth2_refresh_tokens::table)
+                        .values(oauth2::NewRefreshToken {
+                            token: generate_secret().as_str(),
+                            access_token: access_token.token.as_str(),
+                            application_id,
+                        })
+                        .returning(oauth2::RefreshToken::as_returning())
+                        .get_result::<oauth2::RefreshToken>(tx)
+                        .await?;
+
+                    Ok::<_, Error>((access_token, refresh_token))
+                }
+                .scope_boxed()
+            })
+            .await
+            .map_err(|_| ())?;
+
+        Ok(IssuedToken {
+            token: access_token.token,
+            refresh: Some(refresh_token.token),
+            until: grant.until,
+            token_type: TokenType::Bearer,
+        })
+    }
+
+    async fn refresh(&mut self, refresh_token: &str, _grant: Grant) -> Result<RefreshedToken, ()> {
+        let mut db_conn = self.db_pool.get().await.map_err(|_| ())?;
+        let (refresh_token, access_token) = oauth2_refresh_tokens::table
+            .find(refresh_token)
+            .inner_join(oauth2_access_tokens::table)
+            .select(<(oauth2::RefreshToken, oauth2::AccessToken)>::as_select())
+            .get_result::<(oauth2::RefreshToken, oauth2::AccessToken)>(&mut db_conn)
+            .await
+            .map_err(|_| ())?;
+
+        let (access_token, refresh_token) = db_conn
+            .transaction(|tx| {
+                async move {
+                    let new_access_token = diesel::insert_into(oauth2_access_tokens::table)
+                        .values(oauth2::NewAccessToken {
+                            user_id: access_token.user_id,
+                            token: generate_secret().as_str(),
+                            application_id: access_token.application_id,
+                            scopes: access_token.scopes.as_str(),
+                            expired_at: OffsetDateTime::now_utc() + TOKEN_VALID_DURATION,
+                        })
+                        .get_result::<oauth2::AccessToken>(tx)
+                        .await?;
+
+                    let refresh_token = diesel::update(&refresh_token)
+                        .set(
+                            oauth2_refresh_tokens::access_token.eq(new_access_token.token.as_str()),
+                        )
+                        .get_result::<oauth2::RefreshToken>(tx)
+                        .await?;
+
+                    diesel::delete(&access_token).execute(tx).await?;
+
+                    Ok::<_, Error>((new_access_token, refresh_token))
+                }
+                .scope_boxed()
+            })
+            .await
+            .map_err(|_| ())?;
+
+        let until = chrono::NaiveDateTime::from_timestamp_opt(
+            access_token.expired_at.unix_timestamp(),
+            access_token.expired_at.nanosecond(),
+        )
+        .unwrap()
+        .and_utc();
+
+        Ok(RefreshedToken {
+            token: access_token.token,
+            refresh: Some(refresh_token.token),
+            until,
+            token_type: TokenType::Bearer,
+        })
+    }
+
+    async fn recover_token(&mut self, access_token: &str) -> Result<Option<Grant>, ()> {
+        let mut db_conn = self.db_pool.get().await.map_err(|_| ())?;
+        let oauth_data = oauth2_access_tokens::table
+            .find(access_token)
+            .inner_join(oauth2_applications::table)
+            .select(<(oauth2::AccessToken, oauth2::Application)>::as_select())
+            .get_result::<(oauth2::AccessToken, oauth2::Application)>(&mut db_conn)
+            .await
+            .optional()
+            .map_err(|_| ())?;
+
+        let oauth_data = oauth_data.map(|(access_token, app)| {
+            let scope = app.scopes.parse().unwrap();
+            let redirect_uri = app.redirect_uri.parse().unwrap();
+            let until = chrono::NaiveDateTime::from_timestamp_opt(
+                access_token.expired_at.unix_timestamp(),
+                access_token.expired_at.nanosecond(),
+            )
+            .unwrap()
+            .and_utc();
+
+            Grant {
+                owner_id: access_token
+                    .user_id
+                    .as_ref()
+                    .map(ToString::to_string)
+                    .unwrap_or_default(),
+                client_id: app.id.to_string(),
+                scope,
+                redirect_uri,
+                until,
+                extensions: Extensions::default(),
+            }
+        });
+
+        Ok(oauth_data)
+    }
+
+    async fn recover_refresh(&mut self, refresh_token: &str) -> Result<Option<Grant>, ()> {
+        let mut db_conn = self.db_pool.get().await.map_err(|_| ())?;
+        let oauth_data = oauth2_refresh_tokens::table
+            .find(refresh_token)
+            .inner_join(oauth2_access_tokens::table)
+            .inner_join(oauth2_applications::table)
+            .select(<(oauth2::AccessToken, oauth2::Application)>::as_select())
+            .get_result::<(oauth2::AccessToken, oauth2::Application)>(&mut db_conn)
+            .await
+            .optional()
+            .map_err(|_| ())?;
+
+        let oauth_data = oauth_data.map(|(access_token, app)| {
+            let scope = access_token.scopes.parse().unwrap();
+            let redirect_uri = app.redirect_uri.parse().unwrap();
+            let until = chrono::NaiveDateTime::MAX.and_utc();
+
+            Grant {
+                owner_id: access_token
+                    .user_id
+                    .as_ref()
+                    .map(ToString::to_string)
+                    .unwrap_or_default(),
+                client_id: app.id.to_string(),
+                scope,
+                redirect_uri,
+                until,
+                extensions: Extensions::default(),
+            }
+        });
+
+        Ok(oauth_data)
+    }
+}
