+++ kitsune/src/service/oauth2/solicitor.rs
@@ -1,0 +1,139 @@
+use super::OAuthScope;
+use askama::Template;
+use async_trait::async_trait;
+use diesel::{OptionalExtension, QueryDsl};
+use diesel_async::RunQueryDsl;
+use kitsune_db::{model::user::User, schema::oauth2_applications, PgPool};
+use oxide_auth::endpoint::{OAuthError, OwnerConsent, QueryParameter, Solicitation, WebRequest};
+use oxide_auth_async::endpoint::OwnerSolicitor;
+use oxide_auth_axum::{OAuthRequest, OAuthResponse, WebError};
+use std::{borrow::Cow, str::FromStr};
+use strum::EnumMessage;
+use typed_builder::TypedBuilder;
+use uuid::Uuid;
+
+#[derive(Template)]
+#[template(path = "oauth/consent.html")]
+struct ConsentPage<'a> {
+    app_name: &'a str,
+    query: PageQueryParams,
+    scopes: &'a [OAuthScope],
+}
+
+struct PageQueryParams {
+    client_id: String,
+    redirect_uri: String,
+    response_type: String,
+    scope: String,
+    state: Option<String>,
+}
+
+impl PageQueryParams {
+    fn extract(query: &(dyn QueryParameter + 'static)) -> Option<Self> {
+        Some(Self {
+            client_id: query.unique_value("client_id")?.into_owned(),
+            redirect_uri: query.unique_value("redirect_uri")?.into_owned(),
+            response_type: query.unique_value("response_type")?.into_owned(),
+            scope: query.unique_value("scope")?.into_owned(),
+            state: query.unique_value("state").map(Cow::into_owned),
+        })
+    }
+}
+
+#[derive(Clone, TypedBuilder)]
+pub struct OAuthOwnerSolicitor {
+    authenticated_user: User,
+    db_pool: PgPool,
+}
+
+impl OAuthOwnerSolicitor {
+    async fn check_consent(
+        &self,
+        login_consent: Option<&str>,
+        query: PageQueryParams,
+        solicitation: &Solicitation<'_>,
+    ) -> Result<OwnerConsent<OAuthResponse>, WebError> {
+        let consent = match login_consent {
+            Some("accept") => OwnerConsent::Authorized(self.authenticated_user.id.to_string()),
+            Some("deny") => OwnerConsent::Denied,
+            Some(..) | None => {
+                let client_id: Uuid = solicitation
+                    .pre_grant()
+                    .client_id
+                    .parse()
+                    .map_err(|_| WebError::Endpoint(OAuthError::BadRequest))?;
+
+                let mut db_conn = self
+                    .db_pool
+                    .get()
+                    .await
+                    .map_err(|_| WebError::InternalError(None))?;
+
+                let app_name = oauth2_applications::table
+                    .find(client_id)
+                    .select(oauth2_applications::name)
+                    .get_result::<String>(&mut db_conn)
+                    .await
+                    .optional()
+                    .map_err(|_| WebError::InternalError(None))?
+                    .ok_or(WebError::Endpoint(OAuthError::DenySilently))?;
+
+                let scopes = solicitation
+                    .pre_grant()
+                    .scope
+                    .iter()
+                    .map(OAuthScope::from_str)
+                    .collect::<Result<Vec<OAuthScope>, strum::ParseError>>()
+                    .expect("[Bug] Scopes weren't normalised");
+
+                let body = ConsentPage {
+                    app_name: &app_name,
+                    query,
+                    scopes: &scopes,
+                }
+                .render()
+                .map_err(|err| WebError::InternalError(Some(err.to_string())))?;
+
+                OwnerConsent::InProgress(
+                    OAuthResponse::default()
+                        .content_type("text/html")
+                        .unwrap()
+                        .body(&body),
+                )
+            }
+        };
+
+        Ok(consent)
+    }
+}
+
+#[async_trait]
+impl OwnerSolicitor<OAuthRequest> for OAuthOwnerSolicitor {
+    async fn check_consent(
+        &mut self,
+        req: &mut OAuthRequest,
+        solicitation: Solicitation<'_>,
+    ) -> OwnerConsent<OAuthResponse> {
+        let (login_consent, query) = {
+            let query = match req.query() {
+                Ok(query) => query,
+                Err(err) => return OwnerConsent::Error(err),
+            };
+
+            let login_consent = query.unique_value("login_consent").map(Cow::into_owned);
+            let Some(query_params) = PageQueryParams::extract(query.as_ref()) else {
+                return OwnerConsent::Error(WebError::Endpoint(OAuthError::BadRequest));
+            };
+
+            (login_consent, query_params)
+        };
+
+        let result =
+            Self::check_consent(self, login_consent.as_deref(), query, &solicitation).await;
+
+        match result {
+            Ok(consent) => consent,
+            Err(err) => OwnerConsent::Error(err),
+        }
+    }
+}
